Plasma: Introduce PlasmaKVStore
PlasmaKVStore is an experimental KVStore implementation backed by
Plasma. It is very incomplete, and is not intended for general usage.

As such, it is hidden behind the `EP_USE_PLASMA` flag.

To use it, one can build with (for example)

make EXTRA_CMAKE_OPTIONS='-DEP_USE_PLASMA=ON'

As there is no UI involvement, buckets must be created as a normal
Couchbase bucket, and then have the backend changed with the following
diag/eval

 curl -X POST -u Administrator:asdasd http://localhost:9000/diag/eval \
  -d 'ns_bucket:update_bucket_props("bucket_name",
  [{extra_config_string,"backend=plasma"}]).'

Change-Id: Ia4f65af9825968626b0342b95edd3d033563ab64